Lookup Chromedriver version using installed Chrome
This commit updates the chromedriver buildpack to retrieve the correct
chromedriver version based on the installed version of Google Chrome.
During major releases there is a brief window where mismatching versions
can be installed. This change uses the major.minor.patch based on the
[chromedriver docs].

This change requires that heroku-buildpack-google-chrome come first in
the order of buildpacks.

For similar approaches see:

- https://github.com/heroku/heroku-buildpack-chromedriver/pull/28
- https://github.com/heroku/heroku-buildpack-chromedriver/pull/36

[chromedriver docs]: https://chromedriver.chromium.org/downloads/version-selection